+++ packages/app/src/App.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { render } from "react-dom";
 
-import Counter from "./Counter";
+import Counter from "./components/stateful/Counter";
 
 render(<Counter />, document.getElementById("main"));
